+++ extensions/control-plane/client/src/main/java/de/fraunhofer/iosb/client/ClientExtension.java
@@ -31,7 +31,9 @@ import org.eclipse.edc.connector.controlplane.transfer.spi.TransferProcessManage
 import org.eclipse.edc.connector.controlplane.transfer.spi.observe.TransferProcessObservable;
 import org.eclipse.edc.connector.controlplane.transform.odrl.OdrlTransformersFactory;
 import org.eclipse.edc.connector.core.agent.NoOpParticipantIdMapper;
+import org.eclipse.edc.participantcontext.single.spi.SingleParticipantContextSupplier;
 import org.eclipse.edc.runtime.metamodel.annotation.Inject;
+import org.eclipse.edc.spi.EdcException;
 import org.eclipse.edc.spi.system.Hostname;
 import org.eclipse.edc.spi.system.ServiceExtension;
 import org.eclipse.edc.spi.system.ServiceExtensionContext;
@@ -56,6 +58,8 @@ public class ClientExtension implements ServiceExtension {
     @Inject
     private Hostname hostname;
     @Inject
+    private SingleParticipantContextSupplier singleParticipantContextSupplier;
+    @Inject
     private TransferProcessManager transferProcessManager;
     @Inject
     private TransferProcessObservable transferProcessObservable;
@@ -70,11 +74,14 @@ public class ClientExtension implements ServiceExtension {
         var monitor = context.getMonitor().withPrefix("Client");
         var config = context.getConfig("edc.client");
         registerTransformers();
+        var participantContext = singleParticipantContextSupplier.get()
+                .orElseThrow(failure -> new EdcException(failure.getFailureDetail()));
 
         var negotiationController = new NegotiationController(
                 consumerNegotiationManager,
                 contractNegotiationObservable,
                 contractNegotiationStore,
+                participantContext,
                 config);
 
         var transferController = new DataTransferController(
@@ -83,12 +90,14 @@ public class ClientExtension implements ServiceExtension {
                 webService,
                 publicApiManagementService,
                 transferProcessManager,
+                participantContext,
                 transferProcessObservable,
                 hostname);
 
         var policyController = new PolicyController(
                 monitor,
                 catalogService,
+                participantContext,
                 transformer,
                 config);